Added HUD and variant.  Read below
I am taking responsibilities away from the Grid's .draw method since it
is currently doing far too much.  First added is the ModeHUD which draws
the line depending on the mode and draws the mode feedback in the upper
left.  I will expand on HUD later to allow them to be resized.