Layered Drawing Update
Created a layer system for drawing. Layers have a list of sublayers.
When they don't use a click they recieve, they respond false, and their
superlayer sends it to the next sublayer. Layers do not recieve clicks
if the click is not inside of their bounding rectangle so it is
important to set this properly.