Made a small but needed change
Made a small but needed change.  The coding here was slightly out of
order so i put it in the proper order and changed it so that if you want
to change the turn's time, you now only need to change one number
instead of 2.  Also, I made it 3 mins instead of 30 seconds