Added random planet spawner
Added random planet spawner.  This the static Spawner class that will
add planets to the grid at random but will also make sure that the
planets are spawned mirrored over the entire map.  Also, if you request
too many planets to legally be spawned, the static method will time out,
but you will still get the max amount of planets.  This just prevents an
infinite loop from happening when you try to spawn hundreds of planets
on a map that obviously cant hold that many