updated Items and ItemList, new "Overload" Items
The overload items temporarily increase a stat of a ship beyond it's
max, and that stat is decreased turn-by-turn until it returns to its
max, kind of like a temporary "power-up." It could be a power-up found
in the game rather than a purchase-able item, or it could be a
consumable to be bought and used when needed.

It will probably increase a stat a large amount over several turns for a
slightly lower price, to help increase the options for players losing
the game.

I do think I put in all of the functionality necessary for the new
items, but if there's anything I missed just tell me.