Changed InformationPopupLayer
Changed InformationPopupLayer to be a singleton.  This makes popping up
information from anywhere very simple.  I also made some changes in the
code that I'm responsible for in order to make sure it does what its
supposed to with the new changes.  Popups now appear to alert the user
of errors (such as trying to move further than one can). Also, The size
of the popup is dependant on the string input.