Items can now be added to ships via buttons
Some errors are known, such as not being able to properly
deselect/refresh the button selected on the ItemListHUD. Also, strange
exception when right clicking after selecting an item, closing the menu,
then right clicking where the menu used to be. Easiest way to make this
happen is to select an item and then right click the menu 2 times.
Also, there must be a ship docked.